RRW-2: Moved the schematics to the root of the module
